+++ packages/core/lib/FirePHPCore/FirePHP.class.php
@@ -736,7 +736,7 @@ class FirePHP {
      */
     public function fb($Object)
     {
-        if(is_a($this, 'FirePHP_Insight') && method_exists($this, '_logUpgradeClientMessage')) {
+        if($this instanceof FirePHP_Insight && method_exists($this, '_logUpgradeClientMessage')) {
             if(!FirePHP_Insight::$upgradeClientMessageLogged) {    // avoid infinite recursion as _logUpgradeClientMessage() logs a message
                 $this->_logUpgradeClientMessage();
             }