+++ programs/standalone/examples/TestRunner/insight-devcomp/RequestConsole-Groups.php
@@ -11,7 +11,17 @@ $group = $console1->group('Group1')->open();
 
 $console1->log('Group Title');
 $console1->log('Message 1');
+$console1->log('Message 2');
 
-$console2->log('Message 2');
+$console2->log('Message 3');
+
+$group->close();
+
+
+
+$group = $console1->group('Group2', 'Group 2 Title')->open();
+
+$console1->log('Message 1');
+$console1->log('Message 2');
 
 $group->close();