+++ programs/standalone/CHANGELOG.md
@@ -20,6 +20,8 @@ DONE:
   * Enhancement: Provide procedural API via FirePHP/fb.php [5m]
   * Docs: New upgrade page at http://upgrade.firephp.org/ [8h, +misc changes]
   * Bugfix: FirePHP constants when deactivated
+  * Enhancement: Added Zend Framework example (shows if ZF is found)
+  * Bugfix: Use mb_detect_encoding() if available to check encoding of string objects
 
 2010-10-01 - Release Version: 0.0.0master1010012104