+++ programs/standalone/CHANGELOG.md
@@ -23,6 +23,7 @@ DONE:
   * Enhancement: Added Zend Framework example (shows if ZF is found)
   * Bugfix: Use mb_detect_encoding() if available to check encoding of string objects
   * Enhancement: Added Error Reporting levels to FirePHP_Plugin_FirePHP::recordEnvironment [37m]
+  * Bugfix: Use 'instanceof' instead of is_a() [4m]
 
 2010-10-01 - Release Version: 0.0.0master1010012104